Misc changes required to enable c++20 support (#7738)
* Add missing header <bit>

* Add explicit == operators for some Expr derived classes

C++20 considers these comparison functions ambiguous when implemented as
base class comparisons, Implement explicit comparisons for the derived
class and call these from the virtual functions

* Add missing virtual destructors

* Add missing consts to resolve ambiguous overloads

* Change misc.h bit_cast to only use bit_cast if sizes equal

Clang static asserts that both sizes must be equal

* Bitwise operations on different enums is deprecated

Static cast to int to resolve

* Implement explicit constructor for SQLSelectedTablesBuilder

* Always set HAVE_ARROW_4

The test compile fails even though we are well past Arrow 4 resulting in later
build errors. Rather than bother with the test compile just set HAVE_ARROW_4

* Fix C++20 removal of std::result_of<>.

Example error: no member named 'result_of_t' in namespace 'std'
                std::is_void<std::result_of_t<CUSTOM_DELETER(int8_t*)> >::value> >
                             ~~~~~^

* Fix LockMgr/LockMgr.h:179:39: error: expected unqualified-id before ‘const’

* Fix template class constructor declarations

* Fix implicit capture of this by is deprecated

* Remove volatile from UDF avoid_opt_address_geo bool return

This use of volatile has been deprecated, and looking at the code
I'm not sure why it was ever there

* Support bit_cast on big-endian machines.

(cherry picked from commit 95d161c4ad8b6c439ef59d877ff493c09a59b790)

* Move shared::bit_cast outside of #if.

(cherry picked from commit 20a75c852c93738f0b5db30bf0014b52cc46711c)

---------

+++ SQLFrontend/heavysql.cpp
@@ -1046,7 +1046,8 @@ bool get_license_claims(ClientContext& context) {
 std::string hide_sensitive_data_from_connect(const std::string& connect_str) {
   auto result = connect_str;
   boost::regex passwd{R"(^\\c\s+?.+?\s+?.+?\s+?(?<pwd>.+))",
-                      boost::regex_constants::perl | boost::regex::icase};
+                      static_cast<int>(boost::regex_constants::perl) |
+                          static_cast<int>(boost::regex::icase)};
   boost::smatch matches;
   if (boost::regex_search(connect_str, matches, passwd)) {
     result.replace(
